+++ app/src/main/java/campus/tech/kakao/map/data/document/PlaceResponse.kt
@@ -1,4 +1,4 @@
-package campus.tech.kakao.map.dto
+package campus.tech.kakao.map.data.document
 
 data class PlaceResponse(
 	val meta: Meta,
