+++ app/src/main/java/campus/tech/kakao/map/data/searchWord/SearchWordContract.kt
@@ -1,4 +1,4 @@
-package campus.tech.kakao.map.dto
+package campus.tech.kakao.map.data.searchWord
 
 object SearchWordContract {
 	const val DB_NAME = "search_word"
