+++ app/src/main/java/campus/tech/kakao/map/data/searchWord/SearchWordDao.kt
@@ -1,4 +1,4 @@
-package campus.tech.kakao.map.dto
+package campus.tech.kakao.map.data.searchWord
 
 import androidx.room.Dao
 import androidx.room.Insert
