+++ app/src/main/java/campus/tech/kakao/map/view/adapter/DocumentAdapter.kt
@@ -1,4 +1,4 @@
-package campus.tech.kakao.map.adapter
+package campus.tech.kakao.map.view.adapter
 
 import android.view.LayoutInflater
 import android.view.View
@@ -6,7 +6,7 @@ import android.view.ViewGroup
 import androidx.recyclerview.widget.DiffUtil
 import androidx.recyclerview.widget.ListAdapter
 import androidx.recyclerview.widget.RecyclerView
-import campus.tech.kakao.map.dto.Document
+import campus.tech.kakao.map.data.document.Document
 import campus.tech.kakao.map.databinding.PlaceItemBinding
 
 class DocumentAdapter(
