+++ app/src/main/java/campus/tech/kakao/map/viewModel/MainViewModel.kt
@@ -6,14 +6,14 @@ import androidx.lifecycle.AndroidViewModel
 import androidx.lifecycle.LiveData
 import androidx.lifecycle.MutableLiveData
 import androidx.lifecycle.viewModelScope
-import campus.tech.kakao.map.adapter.DocumentAdapter
-import campus.tech.kakao.map.adapter.WordAdapter
+import campus.tech.kakao.map.view.adapter.DocumentAdapter
+import campus.tech.kakao.map.view.adapter.WordAdapter
 import campus.tech.kakao.map.databinding.ActivitySearchBinding
-import campus.tech.kakao.map.dto.Document
-import campus.tech.kakao.map.dto.MapPositionPreferences
-import campus.tech.kakao.map.dto.SearchWord
-import campus.tech.kakao.map.dto.SearchWordDao
-import campus.tech.kakao.map.url.RetrofitData
+import campus.tech.kakao.map.data.document.Document
+import campus.tech.kakao.map.data.mapPosition.MapPositionPreferences
+import campus.tech.kakao.map.data.searchWord.SearchWord
+import campus.tech.kakao.map.data.searchWord.SearchWordDao
+import campus.tech.kakao.map.data.remote.RetrofitData
 import dagger.hilt.android.lifecycle.HiltViewModel
 import kotlinx.coroutines.Dispatchers
 import kotlinx.coroutines.launch
